feat(api/project): move insert_project_into_db and helpers to database module
